update float type to 64 bytes

+++ app/model/gps.go
@@ -2,6 +2,6 @@ package model
 
 type Gps struct {
 	Date []ParsedDate `json:"date"`
-	Long []float32    `json:"long"`
-	Lat  []float32    `json:"lat"`
+	Long []float64    `json:"long"`
+	Lat  []float64    `json:"lat"`
 }
